Rename everything from GitTools to GitSplit
