+++ composer.json
@@ -22,7 +22,7 @@
         "asm/php-ansible": "^2.0",
         "mouf/nodejs-installer": "^1.0",
         "provision-ops/power-process": "^1.3",
-        "devshop/git-tools": "*"
+        "devshop/git-split": "*"
     },
     "require-dev": {
         "composer/composer": "^1.9"
@@ -38,7 +38,7 @@
     "repositories": [
         {
             "type": "path",
-            "url": "src/DevShop/Component/GitTools"
+            "url": "src/DevShop/Component/GitSplit"
         }
     ],
     "bin": [
