+++ src/DevShop/Component/GitSplit/CHANGELOG.md
@@ -1,0 +1,7 @@
+CHANGELOG
+=========
+
+2.0.0
+-----
+
+ * Initial release of the GitSplit component. First ever in the new DevShop 2020 DevOps Framework!
