+++ src/DevShop/Component/GitSplit/Command/GitSplitConsoleCommand.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace DevShop\Component\GitTools\Command;
+namespace DevShop\Component\GitSplit\Command;
 
-use DevShop\Component\GitTools\Splitter;
+use DevShop\Component\GitSplit\Splitter;
 use Symfony\Component\Console\Exception\InvalidOptionException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
