--- src/DevShop/Component/GitTools/CHANGELOG.md
@@ -1,7 +1,0 @@
-CHANGELOG
-=========
-
-2.0.0
------
-
- * Initial release of the GitTools component. First ever in the new DevShop 2020 DevOps Framework!
